added floating alert
