+++ lease-rental/clients/CommunicationManager.swift
