+++ lease-rental/extensions/StringExtension.swift
