+++ lease-rental/leasee/models/Leasee.swift
@@ -1,0 +1,14 @@
+//
+//  Leasee.swift
+//  lease-rental
+//
+//  Created by Ravindu Senevirathna on 3/5/20.
+//  Copyright © 2020 MIHCM. All rights reserved.
+//
+
+import Foundation
+
+struct Leasee: Codable {
+    let id: String
+    let tenant: String
+}
