+++ lease-rental/rental/LeaseRentPaymentsViewModel.swift
@@ -8,72 +8,6 @@
 
 import Foundation
 
-enum PaymentDay: String, Codable {
-    case monday, tuesday, wednesday, thursday, friday
-    
-    var day: Int {
-        switch self {
-        case .monday:
-            return 2
-        case .tuesday:
-            return 3
-        case .wednesday:
-            return 4
-        case .thursday:
-            return 5
-        case .friday:
-            return 6
-        }
-    }
-}
-
-enum PaymentFrequency: String, Codable {
-    case weekly, fortnightly, monthly
-    
-    var value: Int {
-        switch self {
-        case .weekly:
-            return 7
-        case .fortnightly:
-            return 14
-        case .monthly:
-            return 28
-        }
-    }
-}
-
-struct LeaseDetail: Codable {
-    let id, startDate, endDate: String
-    let rent: Double
-    let frequency: PaymentFrequency
-    let paymentDay: PaymentDay
-    var start: Date? {
-        return startDate.toDate()
-    }
-    var end: Date? {
-        return endDate.toDate()
-    }
-    
-    enum CodingKeys: String, CodingKey {
-        case id
-        case startDate = "start_date"
-        case endDate = "end_date"
-        case rent, frequency
-        case paymentDay = "payment_day"
-    }
-}
-
-struct LeaseRental {
-    var from: Date
-    var to: Date
-    var days: Int
-    var rentalPerDay: Double
-    var amount: Double {
-        return Double(days) * rentalPerDay
-    }
-}
-
-
 class LeaseRentPaymentsViewModel: BaseViewModel {
     let service = LeaseService()
     var leaseDetail: Dynamic<LeaseDetail?> = Dynamic(nil)
