+++ lease-rental/rental/views/RentalTableViewCell.xib
